refactor: rename classes and files
